Add both versions to github
Note: Version 1 is a fully functioning todo list for up to 10 todos. It
does not however have the requested input field, it simply adds a new
field and a delete button to delete the field.

Version 2 follows the requested protocol for up to ten todos. To do is
deleted when clicked. The only part which is not functioning is adding
the text from the input field to the todo. I commented out the
functions I built to do this part which I could not get to work in the
allotted time.